second commit

+++ SimpleCalculation/Calculator/Calculation.cs
@@ -8,22 +8,22 @@ namespace Calculator
 {
     public class Calculation
     {
-        public int Addition (int a, int b)
+        public double Addition (double a, double b)
         {
             return a + b;
         }
 
-        public int Subtraction(int a, int b)
+        public double Subtraction(double a, double b)
         {
             return a - b;
         }
 
-        public int Multiplication(int a, int b)
+        public double Multiplication(double a, double b)
         {
             return a * b;
         }
 
-        public int Division(int a, int b)
+        public double Division(double a, double b)
         {
             return a / b;
         }
